[docker] prototype having a proper package.json and yarn.lock
I think the docker example can have its package.json named without
"example-" in front and still be insulated from the rest of the
components forming the mono-repo, so long as it's not added to the
yarn workspace (defined in root package.json).

This makes it easier for the docker example to have a yarn.lock file,
whioch can be used for "yarn audit" / "yarn upgrade", and I think
should permit this exampel application to automatically receive PRs
by Dependabot.

Adjusted Dockerfile and added convenience scripts entry in root package.json
to build the docker application (outside docker), thus updating its
yarn.lock, if/as needed.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>