Make the docker examples into a sub monorepo
We have two docker examples that provide trave viewing capabilities. Each is a Theia
application, built and used in a Docker container. They differ by which extension
provides the trace viewer: one uses the theia-traceviewer extension, the other the
vscode-trace-viewer vscode extension.

With this commit, both examples are made into a sub monorepo, that share a "root" package.json
and yarn.lock, making the two apps manageable together. This is very similar to how the
browser and Electron examples are structured in this repo, sharing the root package.json and
yarn.lock.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>